requirements.txt
python-dotenv
asyncio
langchain
langchain-ollama
streamlit
langchain-openai
langgraph-checkpoint-mongodb
matplotlib